requirements.txt
SQLAlchemy[asyncio]~=2.0.23
python-dotenv~=1.0.0
aiofiles~=23.2.1
asyncmy~=0.2.8
aiohttp~=3.9.0
aiomysql~=0.2.0
python-telegram-bot[ext]~=20.6
pillow~=10.1.0
pixivpy3~=3.7.4
singleton-class-decorator~=1.0.1
b2sdk~=1.27.0
nest_asyncio~=1.5.8